Add watch:test script and changes to better support Windows
Part of this is removing the "mkdirs" script entirely and moving that
functionality into the "clean" script. Also, removing the "clean" script
from many of the calls (except, really "prebuild"). The reason for this is
that calling `shelljs` commands inside scripts running in parralel can
cause errors to be thrown. I'm not sure if it's an issue with
`npm-run-all` or `shelljs`, but I didn't feel like digging too deep on it.